moved datasource credentials in credentials provider
