+++ reference-project/src/main/resources/application.properties
@@ -47,5 +47,4 @@ quarkus.micrometer.binder.http-server.enabled=true
 
 quarkus.datasource.db-kind=postgresql
 quarkus.datasource.jdbc.url=jdbc:postgresql://localhost:5432/demo
-quarkus.datasource.username=demo
-quarkus.datasource.password=demo
+quarkus.datasource.credentials-provider=custom
